+++ app/src/lib.rs
@@ -3,5 +3,5 @@ mod expo;
 mod get;
 mod post;
 pub use expo::expo::Expo;
-pub use get::get_push_receipts;
-pub use post::send_push_notification;
+pub use get::get_push_notification_receipts;
+pub use post::send_push_notifications;
